+++ .idea/libraries/sqlite_jdbc_3_20_1.xml
@@ -1,7 +1,7 @@
 <component name="libraryTable">
   <library name="sqlite-jdbc-3.20.1">
     <CLASSES>
-      <root url="jar://$PROJECT_DIR$/src/hw3/resources/sqlite-jdbc-3.20.1.jar!/" />
+      <root url="jar://$PROJECT_DIR$/src/hw4/resources/sqlite-jdbc-3.20.1.jar!/" />
     </CLASSES>
     <JAVADOC />
     <SOURCES />
