--- src/hw3/client/ClientApp.java
@@ -1,7 +1,0 @@
-package hw3.client;
-
-public class ClientApp {
-    public static void main(String[] args) {
-        new MyClient().setVisible(true);
-    }
-}
