--- src/hw3/client/ServerService.java
@@ -1,21 +1,0 @@
-package hw3.client;
-
-import hw3.server.Message;
-
-import java.io.IOException;
-
-public interface ServerService {
-
-    boolean isConnected();
-
-    void openConnection();
-
-    void closeConnection();
-
-    String authorization(String login, String password) throws IOException;
-
-    void sendMessage(String message);
-
-    Message readMessages();
-
-}
