--- src/hw3/resources/chat_history.txt
@@ -1,2 +1,0 @@
-
-Сервер написал: /clients Dron
