--- src/hw3/server/AuthService.java
@@ -1,9 +1,0 @@
-package hw3.server;
-
-public interface AuthService {
-    void start();
-
-    void stop();
-
-    String getNickByLoginAndPass(String login, String password);
-}
