--- src/hw3/server/BaseAuthService.java
@@ -1,24 +1,0 @@
-package hw3.server;
-
-public class BaseAuthService implements AuthService {
-    private DBManager dbManager;
-
-    public BaseAuthService(DBManager dbManager) {
-        this.dbManager = dbManager;
-    }
-
-    @Override
-    public void start() {
-        System.out.println("Сервис авторизации запущен");
-    }
-
-    @Override
-    public void stop() {
-        System.out.println("Сервис авторизации остановлен");
-    }
-
-    @Override
-    public String getNickByLoginAndPass(String login, String password) {
-        return dbManager.getNickByLoginAndPass(login, password);
-    }
-}
