--- src/hw3/server/ServerApp.java
@@ -1,7 +1,0 @@
-package hw3.server;
-
-public class ServerApp {
-    public static void main(String[] args) {
-        new MyServer();
-    }
-}
