+++ src/hw4/client/ClientApp.java
@@ -1,0 +1,7 @@
+package hw4.client;
+
+public class ClientApp {
+    public static void main(String[] args) {
+        new MyClient().setVisible(true);
+    }
+}
