+++ src/hw4/client/ServerService.java
@@ -1,0 +1,21 @@
+package hw4.client;
+
+import hw4.server.Message;
+
+import java.io.IOException;
+
+public interface ServerService {
+
+    boolean isConnected();
+
+    void openConnection();
+
+    void closeConnection();
+
+    String authorization(String login, String password) throws IOException;
+
+    void sendMessage(String message);
+
+    Message readMessages();
+
+}
