+++ src/hw4/resources/chat_history.txt
@@ -1,0 +1,3 @@
+
+Сервер написал: /clients Niffin 
+Niffin написал: asdsad
