+++ src/hw4/server/AuthService.java
@@ -1,0 +1,9 @@
+package hw4.server;
+
+public interface AuthService {
+    void start();
+
+    void stop();
+
+    String getNickByLoginAndPass(String login, String password);
+}
