+++ src/hw4/server/BaseAuthService.java
@@ -1,0 +1,24 @@
+package hw4.server;
+
+public class BaseAuthService implements AuthService {
+    private DBManager dbManager;
+
+    public BaseAuthService(DBManager dbManager) {
+        this.dbManager = dbManager;
+    }
+
+    @Override
+    public void start() {
+        System.out.println("Сервис авторизации запущен");
+    }
+
+    @Override
+    public void stop() {
+        System.out.println("Сервис авторизации остановлен");
+    }
+
+    @Override
+    public String getNickByLoginAndPass(String login, String password) {
+        return dbManager.getNickByLoginAndPass(login, password);
+    }
+}
