+++ src/hw4/server/ServerApp.java
@@ -1,0 +1,7 @@
+package hw4.server;
+
+public class ServerApp {
+    public static void main(String[] args) {
+        new MyServer();
+    }
+}
